Funkční verze FW s indexováním kanálů le normy SCPI

+++ FW/bp_demo_3/Core/Src/DMM.c
@@ -91,7 +91,7 @@ void DMM_Continous(uint32_t status){
     return;
 }
 
-void DMM_SRate(uint32_t SR){
+void DMM_SRate(uint16_t SR){
     if(ADS121X_DR((uint16_t)SR)){
         
     }else{
